requirements.txt
# Core LLM
openai>=1.50.0
langgraph>=0.2.0

# langgraph peer dependencies (직접 import하지 않지만 버전 호환성을 위해 명시)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.7.0
serpapi>=0.1.5

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
